+++ many-ports-app/src/main/java/pl/gatomek/manyportsapp/random/service/RandomNumberService.java
@@ -1,0 +1,13 @@
+package pl.gatomek.manyportsapp.random.service;
+
+import pl.gatomek.manyportsapp.random.usecase.GenerateRandomNumberUseCase;
+
+import java.util.concurrent.ThreadLocalRandom;
+
+class RandomNumberService implements GenerateRandomNumberUseCase {
+
+    @Override
+    public int generate() {
+        return ThreadLocalRandom.current().nextInt(1000);
+    }
+}
