+++ many-ports-app/src/main/java/pl/gatomek/manyportsapp/uuid/service/UuidService.java
@@ -1,0 +1,13 @@
+package pl.gatomek.manyportsapp.uuid.service;
+
+import pl.gatomek.manyportsapp.uuid.usecase.GenerateUuidUseCase;
+
+import java.util.UUID;
+
+class UuidService implements GenerateUuidUseCase {
+
+    @Override
+    public UUID generate() {
+        return UUID.randomUUID();
+    }
+}
