+++ many-ports-app/src/main/java/pl/gatomek/manyportsapp/uuid/service/UuidServiceConfig.java
@@ -1,0 +1,12 @@
+package pl.gatomek.manyportsapp.uuid.service;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class UuidServiceConfig {
+    @Bean
+    public UuidService uuidService() {
+        return new UuidService();
+    }
+}
