--- many-ports-app/src/main/java/pl/gatomek/manyportsapp/uuid/usecase/UuidService.java
@@ -1,11 +1,0 @@
-package pl.gatomek.manyportsapp.uuid.usecase;
-
-import java.util.UUID;
-
-public class UuidService implements GenerateUuidUseCase {
-
-    @Override
-    public UUID generate() {
-        return UUID.randomUUID();
-    }
-}
